feat(FormReady):Merchant contract & product group & missing fields controls
