+++ apps/web/src/app/[lang]/(main)/(unirefund)/parties/merchants/[partyId]/contracts/new/page.tsx
@@ -1,5 +1,4 @@
 import { auth } from "@repo/utils/auth/next-auth";
-import { isUnauthorized } from "@repo/utils/policies";
 import {
   // getMerchantContractHeadersByMerchantIdApi,
   getRefundTableHeadersAssignablesByMerchantIdApi,
@@ -7,7 +6,13 @@ import {
 import { getMerchantAddressByIdApi } from "src/actions/unirefund/CrmService/actions";
 import ErrorComponent from "src/app/[lang]/(main)/_components/error-component";
 import { getResourceData } from "src/language-data/unirefund/ContractService";
+import { isUnauthorized } from "@repo/utils/policies";
 import MerchantContractHeaderCreateForm from "./components/form";
+import { FormReadyComponent } from "@repo/ui/form-ready";
+import { FileText } from "lucide-react";
+import Link from "next/link";
+import Button from "@repo/ayasofyazilim-ui/molecules/button";
+import { getBaseLink } from "@/utils";
 
 async function getApiRequests(partyId: string) {
   try {
@@ -72,10 +77,28 @@ export default async function Page({
 
   // const biggestContractHeader = otherContractHeadersResponse.data.items?.at(0);
   return (
-    <MerchantContractHeaderCreateForm
-      addressList={addressListResponse.data}
-      languageData={languageData}
-      refundTableHeaders={refundTableHeadersResponse.data}
-    />
+    <FormReadyComponent
+      active={refundTableHeadersResponse.data.length < 1}
+      content={{
+        icon: <FileText className="size-20 text-gray-400" />,
+        title: languageData["Missing.RefundTableHeaders.Title"],
+        message: languageData["Missing.RefundTableHeaders.Message"],
+        action: (
+          <Button asChild className="text-blue-500" variant="link">
+            <Link
+              href={getBaseLink("settings/templates/refund-tables/new", lang)}
+            >
+              {languageData["New"]}
+            </Link>
+          </Button>
+        ),
+      }}
+    >
+      <MerchantContractHeaderCreateForm
+        addressList={addressListResponse.data}
+        languageData={languageData}
+        refundTableHeaders={refundTableHeadersResponse.data}
+      />
+    </FormReadyComponent>
   );
 }
