+++ apps/web/src/language-data/unirefund/CRMService/resources/en.json
@@ -1,6 +1,8 @@
 {
   "Missing.TaxOffices.Title": "Missing Tax Office",
   "Missing.TaxOffices.Message": "There is no tax office in the system. Please create one to continue.",
+  "Missing.ProductGroup.Title": "Missing Product Group",
+  "Missing.ProductGroup.Message": "There is no product group in the system. Please create one to continue.",
 
   "Merchants": "Merchants",
   "Merchants.Description": "You can view stores from here.",
