+++ apps/web/src/language-data/unirefund/ContractService/resources/en.json
@@ -1,6 +1,8 @@
 {
   "Missing.ContractStores.Title": "Missing sub store",
   "Missing.ContractStores.Message": "Please add at least one store to the headquarter.",
+  "Missing.RefundTableHeaders.Title": "Missing refund table headers",
+  "Missing.RefundTableHeaders.Message": "There is no assignable refund table header to this merchant. Please create one to continue.",
   "Contracts.Contract": "Contract",
   "Contracts.RebateSettings": "Rebate settings",
   "Contracts.Stores": "Stores",
