+++ apps/web/src/language-data/unirefund/ContractService/resources/tr.json
@@ -1,6 +1,8 @@
 {
   "Missing.ContractStores.Title": "Eksik alt mağaza",
   "Missing.ContractStores.Message": "Lütfen merkeze en az bir mağaza ekleyin.",
+  "Missing.RefundTableHeaders.Title": "Eksik iade tablosu",
+  "Missing.RefundTableHeaders.Message": "Bu merkezde iade tablosu atanmamış. Lütfen devam etmek için bir iade tablosu oluşturun.",
   "Contracts.Contract": "Sözleşme",
   "Contracts.RebateSettings": "Komisyon ayarları",
   "Contracts.Stores": "Mağazalar",
